refactor: reorganize helper modules and streamline imports
- Moved functions and constants related to progress reporting and embedding from `helpers.progress_constants` and `helpers.exec_summary_progress` to a new `helpers.progress` module for better organization.
- Updated import paths in `analyze.py`, `embedding.py`, `report.py`, and `web.py` to reflect the new structure.
- Removed unused `snippet_lines.py` and consolidated its functionality into `helpers.misc`.
- Enhanced maintainability by centralizing progress-related utilities and improving the clarity of module responsibilities.
