+++ oasis/embedding.py
@@ -20,86 +20,7 @@ from .schemas.function_extract import FunctionExtractResponse
 from .tools import create_cache_dir, logger, chunk_content, parse_input, sanitize_name, open_file
 from pydantic import ValidationError
 
-EMBEDDING_PROGRESS_MIN_INTERVAL_SEC = 2.0
-
-# Minimum spacing between throttle checks when ``min_interval_sec`` is very small or zero:
-# avoids tight loops from hammering embedding hooks while still allowing frequent updates.
-EMBEDDING_THROTTLE_MIN_INTERVAL_FLOOR_SEC = 0.05
-
-# --- Embedding progress hook (throttle) ---------------------------------------
-# Used by EmbeddingManager when reporting file embedding progress to optional
-# callbacks; complements ``oasis.helpers.scan_progress`` for scan-level phases.
-
-
-class EmbeddingProgressThrottle:
-    """Holds throttle state for optional embedding ``(completed, total)`` callbacks."""
-
-    __slots__ = ("_last_emit_mono",)
-
-    def __init__(self) -> None:
-        self._last_emit_mono = 0.0
-
-    def maybe_emit(
-        self,
-        hook: Optional[Callable[[int, int], None]],
-        completed: int,
-        total: int,
-        *,
-        min_interval_sec: float,
-        force: bool = False,
-    ) -> None:
-        """Emit at most once per ``min_interval_sec`` while work is incomplete."""
-        if hook is None:
-            return
-
-        safe_total: int
-        safe_completed: int
-
-        try:
-            total_n = total
-        except (TypeError, ValueError):
-            logger.debug(
-                "EmbeddingProgressThrottle.maybe_emit: invalid total=%r; skipping emit",
-                total,
-            )
-            if not force:
-                return
-            safe_total = 0
-            safe_completed = 0
-        else:
-            total_n = max(0, total_n)
-            if total_n <= 0:
-                logger.debug(
-                    "EmbeddingProgressThrottle.maybe_emit: non-positive total=%r",
-                    total,
-                )
-                if not force:
-                    return
-                safe_total = 0
-                safe_completed = 0
-            else:
-                safe_total = total_n
-                try:
-                    completed_n = completed
-                except (TypeError, ValueError):
-                    completed_n = 0
-                safe_completed = max(0, min(completed_n, safe_total))
-
-        now = time.monotonic()
-
-        if force:
-            self._last_emit_mono = now
-            hook(safe_completed, safe_total)
-            return
-
-        if (
-            safe_completed < safe_total
-            and now - self._last_emit_mono
-            < max(EMBEDDING_THROTTLE_MIN_INTERVAL_FLOOR_SEC, min_interval_sec)
-        ):
-            return
-        self._last_emit_mono = now
-        hook(safe_completed, safe_total)
+from .helpers.embedding import EMBEDDING_PROGRESS_MIN_INTERVAL_SEC, EmbeddingProgressThrottle
 
 
 class EmbeddingManager:
